+++ pkg/apis/ml/v1alpha1/cronjob_status.go
@@ -21,9 +21,10 @@
 package v1alpha1
 
 import (
-	batchApi "k8s.io/api/batch/v1beta1"
+	batchApi "k8s.io/api/batch/v1"
 
 	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
+	"github.com/arangodb/kube-arangodb/pkg/apis/shared"
 )
 
 type ArangoMLCronJobStatus struct {
@@ -34,4 +35,8 @@ type ArangoMLCronJobStatus struct {
 	// +doc/type: batch.CronJobStatus
 	// +doc/link: Kubernetes Documentation|https://godoc.org/k8s.io/api/batch/v1beta1#CronJobStatus
 	batchApi.CronJobStatus `json:",inline"`
+}
+
+func (a *ArangoMLCronJobStatus) Validate() error {
+	return shared.WithErrors(shared.PrefixResourceErrors("spec"))
 }